+++ flask_monitoringdashboard/views/details/__init__.py
@@ -10,9 +10,7 @@
 """
 
 from . import heatmap
-from . import hits_hour
 from . import outliers
-from . import time_hour
 from . import time_user
 from . import time_version
 from . import version_ip